+++ Snake/SnakeClass.cs
@@ -1,0 +1,48 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace Snake
+{
+    class Snake
+    {
+
+        private int xPos;
+        private int yPos;
+        private List<(int, int)> body;
+
+        //static void Main(string[] args)
+        //{
+        public Snake()
+        {
+            body = new List<(int, int)>(3);
+        }
+
+        public static void Eat()
+        {
+
+        }
+
+        public static void LengthenSnake()
+        {
+
+        }
+
+        public static void GetDirection()
+        {
+
+        }
+
+        // Index 1 is the snake head
+        public static void GetSnakePosition()
+        {
+
+        }
+
+        public static void Move()
+        {
+
+        }
+
+        //}
+    }
+}
